Form for parameters.

+++ Build_IT_Web/ClientApp/app/modules/script-interpreter/components/script-form/parameters-form/parameters-form.component.ts
@@ -4,13 +4,14 @@ import { ParameterFilter } from '../../../models/enums/parameter-filter';
 import { Parameter } from '../../../models/interfaces/parameter';
 import { ParameterImpl } from '../../../models/parameterImpl';
 import { ParameterService } from '../../../services/parameter.service';
+import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 
 @Component({
     selector: 'app-parameters-form',
     templateUrl: './parameters-form.component.html',
     styleUrls: ['./parameters-form.component.scss']
 })
-/** parameters-form component*/
+
 export class ParametersFormComponent implements OnInit {
     parameters: Parameter[];
     filteredParameters: Parameter[];
@@ -18,6 +19,7 @@ export class ParametersFormComponent implements OnInit {
 
     scriptId: number;
     editMode: boolean = false;
+    newlyAddedParameter: boolean = false;
     parametersToShow: string = "all";
 
     constructor(private parameterService: ParameterService,
@@ -33,6 +35,46 @@ export class ParametersFormComponent implements OnInit {
             return;
 
         this.getParameters(this.scriptId);
+    }
+
+    drop(event: CdkDragDrop<Parameter[]>): void {
+        moveItemInArray(this.filteredParameters, event.previousIndex, event.currentIndex);
+
+        this.setNumbers(event);
+    }
+
+    private setNumbers(event: CdkDragDrop<Parameter[]>): void {
+        let sortedParameters = this.sortParameters(this.filteredParameters, 'number');
+
+        let addition = this.getAddition();
+
+        sortedParameters[event.previousIndex].number = event.currentIndex + addition;
+
+        if (event.currentIndex < event.previousIndex) {
+            let i = event.previousIndex - 1;
+            for (i; i >= event.currentIndex; i--)
+                sortedParameters[i].number = sortedParameters[i].number + 1;
+        }
+        else if (event.currentIndex > event.previousIndex) {
+            let i = event.currentIndex;
+            for (i; i > event.previousIndex; i--)
+                sortedParameters[i].number = sortedParameters[i].number - 1;
+        }
+    }
+
+    private getAddition() {
+        let addition = 0;
+        if (this.parametersToShow == 'data')
+            addition = 0;
+        if (this.parametersToShow == 'static' || this.parametersToShow == 'calculation') {
+            let parametersFilterCriteria = ParameterFilter['data'];
+            addition += this.parameters.filter(p => (p.context & parametersFilterCriteria) != 0).length;
+        }
+        if (this.parametersToShow == 'calculation') {
+            let parametersFilterCriteria = ParameterFilter['static'];
+            addition += this.parameters.filter(p => (p.context & parametersFilterCriteria) != 0).length;
+        }
+        return addition;
     }
 
     getParameters(id: number) {
@@ -58,16 +100,25 @@ export class ParametersFormComponent implements OnInit {
 
     editParameter(parameter: Parameter) {
         this.editMode = true;
+        this.newlyAddedParameter = false;
         this.newParameter = parameter;
     }
 
     remove(parameterId: number) {
-        this.parameterService.delete(this.scriptId, parameterId)
-            .subscribe((p: Parameter) => {
-                this.parameters = this.parameters.filter(p => p.id != parameterId)
-                this.onParametersToShowChange();
-                console.log("Parameters", p)
-            }, error => console.error(error));
+        if (confirm("Are you sure?"))
+            this.parameterService.delete(this.scriptId, parameterId)
+                .subscribe((p: Parameter) => {
+                    this.parameters = this.parameters.filter(p => p.id != parameterId)
+                    this.onParametersToShowChange();
+                    this.changeNumbering(p.number);
+                    console.log("Parameters", p)
+                }, error => console.error(error));
+    }
+
+    private changeNumbering(number: number) {
+        this.parameters
+            .filter(p => p.number > number)
+            .forEach(p => p.number -= 1);
     }
 
     onCreated(parameter: Parameter) {
@@ -83,8 +134,11 @@ export class ParametersFormComponent implements OnInit {
     }
 
     changeEditMode() {
-        if (!this.editMode)
-            this.newParameter = new ParameterImpl();
+        if (this.editMode) {
+            this.editMode = false;
+            this.newlyAddedParameter = false;
+            this.newParameter = null;
+        }
     }
 
     sortParameters(parameters: Parameter[], prop: string) {
@@ -95,33 +149,14 @@ export class ParametersFormComponent implements OnInit {
                         -1);
     }
 
-    moveUp(parameter: Parameter) {
-        let sortedParameters = this.sortParameters(this.parameters, 'number');
-        let currentIndex = sortedParameters.indexOf(parameter);
-        if (currentIndex === sortedParameters.length - 1)
-            return;
-
-        let tempNumber = parameter.number;
-        parameter.number = sortedParameters[currentIndex + 1].number;
-        sortedParameters[currentIndex + 1].number = tempNumber;
-
-        this.saveParameters();
+    addNewParameter(): void {
+        this.editMode = true;
+        this.newlyAddedParameter = true;
+        this.newParameter = new ParameterImpl();
+        this.newParameter.number = Math.max.apply(Math, this.parameters.map(function (p) { return p.number; })) + 1;
     }
 
-    moveDown(parameter: Parameter) {
-        let sortedParameters = this.sortParameters(this.parameters, 'number');
-        let currentIndex = sortedParameters.indexOf(parameter);
-        if (currentIndex === 0)
-            return;
-
-        let tempNumber = parameter.number;
-        parameter.number = sortedParameters[currentIndex - 1].number;
-        sortedParameters[currentIndex - 1].number = tempNumber;
-
-        this.saveParameters();
-    }
-
-    private saveParameters() {
+    saveParameters() {
         this.parameters.forEach(p => {
             this.parameterService.update(this.scriptId, p)
                 .subscribe((p: Parameter) => {
